==> 3-clean_final3.ipynb <==
from google.colab import drive
# Mount google drive
drive.mount('/content/drive')
%cd /content/drive/MyDrive/Lab/ODoH-ML/missouri-third-round-24nov/2/

%ls

import pandas as pd

# Load dataset file
df_cleaned = pd.read_csv("final3.csv")
df_cleaned.columns

df_cleaned2 = df_cleaned[[
                          'Name_x' , 
                          'groups',

                          'Updated Proxy to Target Time since previous frame in this TCP stream',
                          'Updated Target to Proxy - Time since previous frame in this TCP stream',

                          'Target to Proxy HTTP-Content-Length',
                          'Proxy to Target HTTP-Content-Length',

                          'Resolver in Target-Resolver',

                          'Time Target to Resolver', 
                          'Time Resolver to Target',

                          'DNS Length Resolver to Target' , 
                          'DNS Length Target to Resolver' 
                          ]]

df_cleaned2.head(5)

df_cleaned2 = df_cleaned2.rename({
    
                          'Name_x': 'sub_query' , 
                          'groups' : 'query',


                          'Updated Proxy to Target Time since previous frame in this TCP stream': 'Input-U-Time',
                          'Updated Target to Proxy - Time since previous frame in this TCP stream': 'Input-D-Time',

                          'Proxy to Target HTTP-Content-Length': 'Input-U-Size',
                          'Target to Proxy HTTP-Content-Length': 'Input-D-Size',

                          'Time Target to Resolver': 'Out-U-Time', 
                          'Time Resolver to Target': 'Out-D-Time',

                          'DNS Length Target to Resolver' : 'Out-U-Size',
                          'DNS Length Resolver to Target' : 'Out-D-Size' 

                    }, axis=1) 

df_cleaned2.head(15)

# saving the dataframe
df_cleaned2.to_csv('final3_cleaned.csv')

df_cleaned2.columns

df_cleaned3 = df_cleaned2
df_cleaned3.insert(1, 'Count', -1)
df_cleaned3.head(30)

# df_cleaned4 = df_cleaned3.groupby(['groups']).mean()
df_cleaned3 = df_cleaned3.groupby(['query']).agg({
                                                   'Input-U-Time': 'sum',
                                                   'Input-D-Time':'sum',

                                                   'Input-D-Size':'sum',
                                                   'Input-D-Size':'sum',

                                                   'Out-U-Time' : 'sum',
                                                   'Out-D-Time' : 'sum',

                                                   'Out-U-Size': 'sum',
                                                   'Out-D-Size': 'sum',

                                                   'Count' : 'count',

                                             }).reset_index()

df_cleaned3

df_cleaned3 = df_cleaned3.rename({
                                                   'Input-U-Time': 'sum Input-U-Time',
                                                   'Input-D-Time':'sum Input-D-Time',

                                                   'Input-D-Size':'sum Input-D-Size',
                                                   'Input-D-Size':'sum Input-D-Size',

                                                   'Out-U-Time' : 'sum Out-U-Time',
                                                   'Out-D-Time' : 'sum Out-D-Time',

                                                   'Out-U-Size': 'sum Out-U-Size',
                                                   'Out-D-Size': 'sum Out-D-Size',

                                                   'Count' : 'count',

                    }, axis=1) 

# saving the dataframe
df_cleaned3.to_csv('final3_cleaned_grouped.csv')
