==> 4-final3_raw_all_adding_query_id.ipynb <==
from google.colab import drive
drive.mount('/content/drive')
%cd /content/drive/MyDrive/Lab/ODoH-ML/missouri-third-round-24nov/2/

%ls

import pandas as pd
# ds_1 = pd.read_csv("1/final3_cleaned.csv")
# ds_2 = pd.read_csv("2/final3_cleaned.csv")
# ds_3 = pd.read_csv("3/final3_cleaned.csv")
# ds_4 = pd.read_csv("4/final3_cleaned.csv")
# ds_5 = pd.read_csv("5/final3_cleaned.csv")
# ds_6 = pd.read_csv("6/final3_cleaned.csv")

ds_6 = pd.read_csv("final3_cleaned.csv")

# ds_1.info()
# ds_2.info()
# ds_3.info()
# ds_4.info()
# ds_5.info()

ds_temp = [ds_6]
  
ds = pd.concat(ds_temp,ignore_index=True)
ds
ds = ds.iloc[: , 1:]
ds.info()

ds.head(30)

ds.loc[1,'sub_query']

for i in range(len(ds['sub_query'])):
  if(ds.loc[i, 'query'] == 'client-mo.test3.ch-geni-net.instageni.rnet.missouri.edu'):
    ds.loc[i,'query'] = ds.loc[i,'sub_query']

for i in range(len(ds['sub_query'])):
  if(ds.loc[i, 'query'] == 'client-mo.test3.ch-geni-net.instageni.rnet.missouri.edu'):
    ds = ds.drop(index=[i])

ds = ds.reset_index(drop=True)
ds

ds.head(5)

%ls

ds_updated = ds
ds_updated.head(10)

# from sklearn import preprocessing

# le = preprocessing.LabelEncoder()
# le.fit(ds[['query']])

# ds_updated['query_id'] = le.transform(ds[['query']])

import pandas as pd

# Load dataset file
domains_alphabet_unique = list(ds['query'].unique())
domains_alphabet = ds['query']

domains_number = []

for i in range(len(domains_alphabet)):
  if(domains_alphabet[i] in domains_alphabet_unique):
   key = domains_alphabet[i]
   domains_number.append(domains_alphabet_unique.index(key))
  # print(domains_alphabet_unique.index(key))


# domains_number
print(len(domains_number))
print((domains_alphabet_unique))

ds_updated['query_id'] = domains_number

ds_updated.head(5)

ds_updated.to_csv('MEREGED_RAW_FINAL.csv')
ds_updated.info()
