==> 6-Grouping-FINAL-MERGED.ipynb <==
from google.colab import drive
# Mount google drive
drive.mount('/content/drive')
%cd /content/drive/MyDrive/Lab/ODoH-ML/missouri-third-round-24nov/

%ls

import pandas as pd
ds_1_grouped = pd.read_csv("1/MEREGED_RAW_FINAL_GROUPED.csv")
ds_2_grouped = pd.read_csv("2/MEREGED_RAW_FINAL_GROUPED.csv")
# ds_3_grouped = pd.read_csv("3/MEREGED_RAW_FINAL_GROUPED.csv")
# ds_4_grouped = pd.read_csv("4/MEREGED_RAW_FINAL_GROUPED.csv")
# ds_5_grouped = pd.read_csv("5/MEREGED_RAW_FINAL_GROUPED.csv")
# ds_6_grouped = pd.read_csv("6/MEREGED_RAW_FINAL_GROUPED.csv")


ds_1 = pd.read_csv("1/MEREGED_RAW_FINAL.csv")
ds_2 = pd.read_csv("2/MEREGED_RAW_FINAL.csv")
# ds_3 = pd.read_csv("3/MEREGED_RAW_FINAL.csv")
# ds_4 = pd.read_csv("4/MEREGED_RAW_FINAL.csv")
# ds_5 = pd.read_csv("5/MEREGED_RAW_FINAL.csv")
# ds_6 = pd.read_csv("6/MEREGED_RAW_FINAL.csv")

ds_2_grouped.info()

ds_temp_grouped = [ds_1_grouped, ds_2_grouped]
ds_temp = [ds_1, ds_2]

ds_grouped = pd.concat(ds_temp_grouped,ignore_index=True)
ds = pd.concat(ds_temp,ignore_index=True)

ds.head(5)

%ls /content/drive/MyDrive/Lab/ODoH-ML/NOVEMBER-100-TOP-ALEXA/

ds_grouped.info()

print(len(ds_grouped['query'].unique()))
print(len(ds['query'].unique()))

ds_grouped.to_csv('/content/drive/MyDrive/Lab/ODoH-ML/NOVEMBER-100-TOP-ALEXA/missouri/NOV-25-ALL-GROUPED-MEREGED-missouri-third.csv')
ds.to_csv('/content/drive/MyDrive/Lab/ODoH-ML/NOVEMBER-100-TOP-ALEXA/missouri/NOV-25-ALL-MEREGED-missouri-third.csv')

%ls /content/drive/MyDrive/Lab/ODoH-ML/NOVEMBER-100-TOP-ALEXA/
