==> 7-Last.ipynb <==
from google.colab import drive
drive.mount('/content/drive')
%cd /content/drive/MyDrive/Lab/ODoH-ML/NOVEMBER-100-TOP-ALEXA/

%ls

import pandas as pd

CATEGORY_PATH = "/content/drive/MyDrive/Lab/ODoH-ML/NOV-10-wisconsin-top-100-alexa/category.csv"

top_100_domains = pd.read_csv(CATEGORY_PATH)
top_100_domains_list = list(top_100_domains['website'])

import pandas as pd

##Georgia
ds_1g = pd.read_csv("georgia/NOV-11-ALL-GROUPED-MEREGED-Georgia.csv")
ds_2g = pd.read_csv("georgia/NOV-22-ALL-GROUPED-MEREGED-georgia-repeat2.csv")
ds_3g = pd.read_csv("georgia/NOV-22-ALL-GROUPED-MEREGED-georgia-repeat3.csv")
##Seattle
ds_4g = pd.read_csv("seattle/NOV-9-ALL-GROUPED-MEREGED-seattle.csv")
ds_5g = pd.read_csv("seattle/NOV-23-ALL-GROUPED-MEREGED-seattle-repeat3.csv")
ds_6g = pd.read_csv("seattle/NOV-16-ALL-GROUPED-MEREGED-seattle-repeat2.csv")
##Wisconsin
ds_7g = pd.read_csv("wisconsin/NOV-5-ALL-GROUPED-MEREGED-wisconsin.csv")
ds_8g = pd.read_csv("wisconsin/NOV-10-ALL-GROUPED-MEREGED-wisconsin.csv")
ds_9g = pd.read_csv("wisconsin/Bunch1-NOV-10-ALL-GROUPED-MEREGED-wisconsin.csv")
##Texas
ds_10g = pd.read_csv("texas/NOV-19-ALL-GROUPED-MEREGED-texas-repeat4.csv")
ds_11g = pd.read_csv("texas/NOV-18-ALL-GROUPED-MEREGED-texas-repeat3.csv")
ds_12g = pd.read_csv("texas/NOV-17-ALL-GROUPED-MEREGED-texas-repeat2.csv")
ds_13g = pd.read_csv("texas/NOV-13-ALL-GROUPED-MEREGED-Texas.csv")
## missouri
ds_14g = pd.read_csv("missouri/NOV-23-ALL-GROUPED-MEREGED-missouri-first.csv")
ds_15g = pd.read_csv("missouri/NOV-24-ALL-GROUPED-MEREGED-missouri-second.csv")
ds_16g = pd.read_csv("missouri/NOV-25-ALL-GROUPED-MEREGED-missouri-third.csv")
############
############
##Georgia
ds_1 = pd.read_csv("georgia/NOV-11-ALL-MEREGED-Georgia.csv")
ds_2 = pd.read_csv("georgia/NOV-22-ALL-MEREGED-georgia-repeat2.csv")
ds_3 = pd.read_csv("georgia/NOV-22-ALL-MEREGED-georgia-repeat3.csv")
##Seattle
ds_4 = pd.read_csv("seattle/NOV-9-ALL-MEREGED-seattle.csv")
ds_5 = pd.read_csv("seattle/NOV-16-ALL-MEREGED-Seattle-repeat2.csv")
ds_6 = pd.read_csv("seattle/NOV-23-ALL-MEREGED-seattle-repeat3.csv")
##Wisconsin
ds_7 = pd.read_csv("wisconsin/NOV5-ALL-MEREGED-wisconsin.csv")
ds_8 = pd.read_csv("wisconsin/NOV-10-ALL-MEREGED-wisconsin.csv")
ds_9 = pd.read_csv("wisconsin/Bunch1-NOV-10-ALL-MEREGED-wisconsin.csv")
##Texas
ds_10 = pd.read_csv("texas/NOV-13-ALL-MEREGED-Texas.csv")
ds_11 = pd.read_csv("texas/NOV-17-ALL-MEREGED-texas-repeat2.csv")
ds_12 = pd.read_csv("texas/NOV-18-ALL-MEREGED-texas-repeat3.csv")
ds_13 = pd.read_csv("texas/NOV-19-ALL-MEREGED-texas-repeat4.csv")
## missouri
ds_14 = pd.read_csv("missouri/NOV-23-ALL-MEREGED-missouri-first.csv")
ds_15 = pd.read_csv("missouri/NOV-24-ALL-MEREGED-missouri-second.csv")
ds_16 = pd.read_csv("missouri/NOV-25-ALL-MEREGED-missouri-third.csv")

# ds_5.loc[ds_5['query'] == "patreon.com"]
ds_3g.info()

# ds_2.loc[ds_2['query'] == "patreon.com"]

new_query1g = []
new_query2g = []
new_query3g = []
new_query4g = []
new_query5g = []
new_query6g = []
new_query7g = []
new_query8g = []
new_query9g = []
new_query10g = []
new_query11g = []
new_query12g = []
new_query13g = []
new_query14g = []
new_query15g = []
new_query16g = []
#####################
new_query1 = []
new_query2 = []
new_query3 = []
new_query4 = []
new_query5 = []
new_query6 = []
new_query7 = []
new_query8 = []
new_query9 = []
new_query10 = []
new_query11 = []
new_query12 = []
new_query13 = []
new_query14 = []
new_query15 = []
new_query16 = []


for i in range(len(ds_1g)):
  new_query1g.append(ds_1g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_2g)):
  new_query2g.append(ds_2g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_3g)):
  new_query3g.append(ds_3g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_4g)):
  new_query4g.append(ds_4g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_5g)):
  new_query5g.append(ds_5g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_6g)):
  new_query6g.append(ds_6g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_7g)):
  new_query7g.append(ds_7g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_8g)):
  new_query8g.append(ds_8g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_9g)):
  new_query9g.append(ds_9g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_10g)):
  new_query10g.append(ds_10g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_11g)):
  new_query11g.append(ds_11g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_12g)):
  new_query12g.append(ds_12g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_13g)):
  new_query13g.append(ds_13g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_14g)):
  new_query14g.append(ds_14g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_15g)):
  new_query15g.append(ds_15g.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_16g)):
  new_query16g.append(ds_16g.iloc[i]["query"].replace('www.', ''))
##########################
for i in range(len(ds_1)):
  new_query1.append(ds_1.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_2)):
  new_query2.append(ds_2.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_3)):    
  new_query3.append(ds_3.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_4)):
  new_query4.append(ds_4.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_5)):
  new_query5.append(ds_5.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_6)):
  new_query6.append(ds_6.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_7)):
  new_query7.append(ds_7.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_8)):
  new_query8.append(ds_8.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_9)):
  new_query9.append(ds_9.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_10)):
  new_query10.append(ds_10.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_11)):
  new_query11.append(ds_11.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_12)):
  new_query12.append(ds_12.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_13)):
  new_query13.append(ds_13.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_14)):
  new_query14.append(ds_14.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_15)):
  new_query15.append(ds_15.iloc[i]["query"].replace('www.', ''))

for i in range(len(ds_16)):
  new_query16.append(ds_16.iloc[i]["query"].replace('www.', ''))

ds_1g["query_2"] = new_query1g
ds_2g["query_2"] = new_query2g
ds_3g["query_2"] = new_query3g
ds_4g["query_2"] = new_query4g
ds_5g["query_2"] = new_query5g
ds_6g["query_2"] = new_query6g
ds_7g["query_2"] = new_query7g
ds_8g["query_2"] = new_query8g
ds_9g["query_2"] = new_query9g
ds_10g["query_2"] = new_query10g
ds_11g["query_2"] = new_query11g
ds_12g["query_2"] = new_query12g
ds_13g["query_2"] = new_query13g
ds_14g["query_2"] = new_query14g
ds_15g["query_2"] = new_query15g
ds_16g["query_2"] = new_query16g
###########
ds_1["query_2"] = new_query1
ds_2["query_2"] = new_query2
ds_3["query_2"] = new_query3
ds_4["query_2"] = new_query4
ds_5["query_2"] = new_query5
ds_6["query_2"] = new_query6
ds_7["query_2"] = new_query7
ds_8["query_2"] = new_query8
ds_9["query_2"] = new_query9
ds_10["query_2"] = new_query10
ds_11["query_2"] = new_query11
ds_12["query_2"] = new_query12
ds_13["query_2"] = new_query13
ds_14["query_2"] = new_query14
ds_15["query_2"] = new_query15
ds_16["query_2"] = new_query16

li1 = ds_14g["query_2"]
li2 = top_100_domains_list
 
set_dif = set(li1).symmetric_difference(set(li2))
list_1_2_dif = list(set_dif)
print(list_1_2_dif)

for i in range(len(ds_1g['query_2'])):
  if(ds_1g.loc[i, 'query_2'] not in top_100_domains_list):
    # print(ds_1.loc[i, 'query_2'])
    ds_1g = ds_1g.drop(index=[i])

for i in range(len(ds_2g['query_2'])):
  if(ds_2g.loc[i, 'query_2'] not in top_100_domains_list):
    # print(ds_2g.loc[i, 'query_2'])
    ds_2g = ds_2g.drop(index=[i])

for i in range(len(ds_3g['query_2'])):
  if(ds_3g.loc[i, 'query_2'] not in top_100_domains_list):
    # print(ds_3g.loc[i, 'query_2'])
    ds_3g = ds_3g.drop(index=[i])
  
for i in range(len(ds_4g['query_2'])):
  if(ds_4g.loc[i, 'query_2'] not in top_100_domains_list):
    # print(ds_4g.loc[i, 'query_2'])
    ds_4g = ds_4g.drop(index=[i])

for i in range(len(ds_5g['query_2'])):
  if(ds_5g.loc[i, 'query_2'] not in top_100_domains_list):
    # print(ds_5g.loc[i, 'query_2'])
    ds_5g = ds_5g.drop(index=[i])

for i in range(len(ds_6g['query_2'])):
  if(ds_6g.loc[i, 'query_2'] not in top_100_domains_list):
    # print(ds_6g.loc[i, 'query_2'])
    ds_6g = ds_6g.drop(index=[i])

for i in range(len(ds_7g['query_2'])):
  if(ds_7g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_7g.loc[i, 'query_2'])
    ds_7g = ds_7g.drop(index=[i])

for i in range(len(ds_8g['query_2'])):
  if(ds_8g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_8g.loc[i, 'query_2'])
    ds_8g = ds_8g.drop(index=[i])

for i in range(len(ds_9g['query_2'])):
  if(ds_9g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_9g.loc[i, 'query_2'])
    ds_9g = ds_9g.drop(index=[i])

for i in range(len(ds_10g['query_2'])):
  if(ds_10g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_10g.loc[i, 'query_2'])
    ds_10g = ds_10g.drop(index=[i])

for i in range(len(ds_11g['query_2'])):
  if(ds_11g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_11g.loc[i, 'query_2'])
    ds_11g = ds_11g.drop(index=[i])

for i in range(len(ds_12g['query_2'])):
  if(ds_12g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_12g.loc[i, 'query_2'])
    ds_12g = ds_12g.drop(index=[i])

for i in range(len(ds_13g['query_2'])):
  if(ds_13g.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_13g.loc[i, 'query_2'])
    ds_13g = ds_13g.drop(index=[i])
##################



for i in range(len(ds_1['query_2'])):
  if(ds_1.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_1.loc[i, 'query_2'])
    ds_1 = ds_1.drop(index=[i])

for i in range(len(ds_2['query_2'])):
  if(ds_2.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_2.loc[i, 'query_2'])
    ds_2 = ds_2.drop(index=[i])

for i in range(len(ds_3['query_2'])):
  if(ds_3.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_3.loc[i, 'query_2'])
    ds_3 = ds_3.drop(index=[i])

for i in range(len(ds_4['query_2'])):
  if(ds_4.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_4.loc[i, 'query_2'])
    ds_4 = ds_4.drop(index=[i])

for i in range(len(ds_5['query_2'])):
  if(ds_5.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_5.loc[i, 'query_2'])
    ds_5 = ds_5.drop(index=[i])

for i in range(len(ds_6['query_2'])):
  if(ds_6.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_6.loc[i, 'query_2'])
    ds_6 = ds_6.drop(index=[i])

for i in range(len(ds_7['query_2'])):
  if(ds_7.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_7.loc[i, 'query_2'])
    ds_7 = ds_7.drop(index=[i])

for i in range(len(ds_8['query_2'])):
  if(ds_8.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_8.loc[i, 'query_2'])
    ds_8 = ds_8.drop(index=[i])

for i in range(len(ds_9['query_2'])):
  if(ds_9.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_9.loc[i, 'query_2'])
    ds_9 = ds_9.drop(index=[i])

for i in range(len(ds_10['query_2'])):
  if(ds_10.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_10.loc[i, 'query_2'])
    ds_10 = ds_10.drop(index=[i])

for i in range(len(ds_11['query_2'])):
  if(ds_11.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_11.loc[i, 'query_2'])
    ds_11 = ds_11.drop(index=[i])

for i in range(len(ds_12['query_2'])):
  if(ds_12.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_12.loc[i, 'query_2'])
    ds_12 = ds_12.drop(index=[i])

for i in range(len(ds_13['query_2'])):
  if(ds_13.loc[i, 'query_2'] not in top_100_domains_list):
    print(ds_13.loc[i, 'query_2'])
    ds_13 = ds_13.drop(index=[i])

################
ds_1g = ds_1g.reset_index(drop=True)
ds_2g = ds_2g.reset_index(drop=True)
ds_3g = ds_3g.reset_index(drop=True)
ds_4g = ds_4g.reset_index(drop=True)
ds_5g = ds_5g.reset_index(drop=True)
ds_6g = ds_6g.reset_index(drop=True)
ds_7g = ds_7g.reset_index(drop=True)
ds_8g = ds_8g.reset_index(drop=True)
ds_9g = ds_9g.reset_index(drop=True)
ds_10g = ds_10g.reset_index(drop=True)
ds_11g = ds_11g.reset_index(drop=True)
ds_12g = ds_12g.reset_index(drop=True)
ds_13g = ds_13g.reset_index(drop=True)
###############
ds_1 = ds_1.reset_index(drop=True)
ds_2 = ds_2.reset_index(drop=True)
ds_3 = ds_3.reset_index(drop=True)
ds_4 = ds_4.reset_index(drop=True)
ds_5 = ds_5.reset_index(drop=True)
ds_6 = ds_6.reset_index(drop=True)
ds_7 = ds_7.reset_index(drop=True)
ds_8 = ds_8.reset_index(drop=True)
ds_9 = ds_9.reset_index(drop=True)
ds_10 = ds_10.reset_index(drop=True)
ds_11 = ds_11.reset_index(drop=True)
ds_12 = ds_12.reset_index(drop=True)
ds_13 = ds_13.reset_index(drop=True)

len(ds_13g['query_2'].unique())

# li3 = ds_3["query_2"]
# li4 = ds_4["query_2"]
 
# set_dif = set(li3).symmetric_difference(set(li4))
# list_3_4_dif = list(set_dif)
# print(list_3_4_dif)

# li5 = ds_5["query_2"]
# li6 = ds_6["query_2"]
 
# set_dif = set(li5).symmetric_difference(set(li6))
# list_5_6_dif = list(set_dif)
# print(list_5_6_dif)

# for i in range(len(ds_1['query_2'])):
#   if(ds_1.loc[i, 'query_2'] in list_1_2_dif):
#     ds_1 = ds_1.drop(index=[i])

# for i in range(len(ds_2['query_2'])):
#   if(ds_2.loc[i, 'query_2'] in list_1_2_dif):
#     ds_2 = ds_2.drop(index=[i])
# ############
# for i in range(len(ds_3['query_2'])):
#   if(ds_3.loc[i, 'query_2'] in list_3_4_dif):
#     ds_3 = ds_3.drop(index=[i])

# for i in range(len(ds_4['query_2'])):
#   if(ds_4.loc[i, 'query_2'] in list_3_4_dif):
#     ds_4 = ds_4.drop(index=[i])
# ############
# for i in range(len(ds_5['query_2'])):
#   if(ds_5.loc[i, 'query_2'] in list_5_6_dif):
#     ds_5 = ds_5.drop(index=[i])

# for i in range(len(ds_6['query_2'])):
#   if(ds_6.loc[i, 'query_2'] in list_5_6_dif):
#     ds_6 = ds_6.drop(index=[i])
# ############

# ds_1 = ds_1.reset_index(drop=True)
# ds_2 = ds_2.reset_index(drop=True)
# ds_3 = ds_3.reset_index(drop=True)
# ds_4 = ds_4.reset_index(drop=True)
# ds_5 = ds_5.reset_index(drop=True)
# ds_6 = ds_6.reset_index(drop=True)

# ds_4.info()

ds_final_grouped = [ds_1g, ds_2g, ds_3g, ds_4g, ds_5g, ds_6g, ds_7g, ds_8g, ds_9g, ds_10g, ds_11g, ds_12g, ds_13g]
ds_final = [ds_1, ds_2, ds_3, ds_4, ds_5, ds_6, ds_7, ds_8, ds_9, ds_10, ds_11, ds_12, ds_13]

ds_final_grouped = pd.concat(ds_final_grouped,ignore_index=True)
ds_final = pd.concat(ds_final,ignore_index=True)

ds_final_grouped.info()

from sklearn import preprocessing

le = preprocessing.LabelEncoder()
le.fit(ds_final[['query_2']])

ds_final_grouped['categorical_label'] = le.transform(ds_final_grouped[['query_2']])
ds_final['categorical_label'] = le.transform(ds_final[['query_2']])

ds_final.info()

len(ds_final_grouped['categorical_label'].unique())

ds_final_grouped.to_csv('THE-LAST-GROUPED.csv')
ds_final.to_csv('THE-LAST.csv')
